+++ CHANGES.rst
@@ -44,6 +44,11 @@
 
 - Stop checking the status of ``readCurrent`` OIDs twice.
 
+- Make the gevent MySQL driver yield more frequently while getting
+  large result sets. Previously it would block in C to read the entire
+  result set. Now it yields according to the cursor's ``arraysize``.
+  See :issue:`315`.
+
 3.0a7 (2019-08-07)
 ==================